requirements.txt
tensorflow>=2.10.0
numpy~=2.0.2
Pillow~=11.0.0
matplotlib~=3.9.2
deeplake